requirements.txt
tensorflow-cpu==2.15.0
streamlit
numpy
pandas
matplotlib
scikit-learn
